feat : 게시판 세션,css수정

+++ src/main/java/com/kb04/starroad/Controller/BoardController3.java
@@ -3,18 +3,20 @@ package com.kb04.starroad.Controller;
 import com.kb04.starroad.Dto.board.BoardResponseDto;
 import com.kb04.starroad.Dto.board.CommentDto;
 import com.kb04.starroad.Entity.Board;
+import com.kb04.starroad.Entity.Member;
 import com.kb04.starroad.Service.BoardService2;
 import com.kb04.starroad.Service.CommentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpSession;
 import java.util.Base64;
 import java.util.List;
 import java.util.Optional;
 
 @RestController
-public class BoardController3 {
+public class    BoardController3 {
 
     @Autowired
     private BoardService2 boardService2;
@@ -24,10 +26,22 @@ public class BoardController3 {
 
 
     @GetMapping("/starroad/board/delete")
-    public ModelAndView deleteBoard(@RequestParam Integer no) {
+    public ModelAndView deleteBoard(@RequestParam Integer no, HttpSession session) {
+        ModelAndView mav = new ModelAndView();
 
-        boardService2.deleteBoard(no);
-        ModelAndView mav = new ModelAndView("redirect:/starroad/board/main");
+        Member currentUser = (Member) session.getAttribute("currentUser");
+        String currentUserId = currentUser.getId();
+
+        if (boardService2.canDelete(no, currentUserId)) {
+            // 현재 사용자가 게시글 삭제 가능한 경우
+            boardService2.deleteBoard(no);
+            mav.setViewName("redirect:/starroad/board/main");
+
+        } else{
+            // 현재 사용자가 게시글 삭제 불가능한 경우 or 게시글 존재하지 않는 경우
+            mav.setViewName("board/deleteError");
+        }
+
         return mav;
     }
     @GetMapping("/starroad/board/detail")
